# backend/gwisi40server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "ISI 4.0 server API",
        "title": "ISI 4.0 server API",
        "contact": {
            "name": "ICTS",
            "url": "http://www.grupoicts.com.br"
        },
        "license": {
            "name": "Proprietary"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/protocol": {
            "post": {
                "description": "Creates a protocol",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Protocols"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Add protocol",
                        "name": "type",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/modelsswagger.Protocol"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Protocol"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "Missing"
                        }
                    },
                    "401": {
                        "description": "Request not authorized"
                    }
                }
            }
        },
        "/api/v1/protocol/{id}": {
            "get": {
                "description": "Get protocol by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Protocols"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Protocol ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Protocol"
                        }
                    },
                    "401": {
                        "description": "Request not authorized"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "No"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a protocol",
                "tags": [
                    "Protocols"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Protocol ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Protocol"
                        }
                    },
                    "401": {
                        "description": "Request not authorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "type": "Cannot"
                        }
                    }
                }
            }
        },
        "/api/v1/protocol/{protocol}": {
            "get": {
                "description": "Get protocol by name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Protocols"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Protocol ID",
                        "name": "protocol",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Protocol"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "No"
                        }
                    }
                }
            },
            "patch": {
                "description": "Updates a protocol",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Protocols"
                ],
                "parameters": [
                    {
                        "description": "Add protocol",
                        "name": "type",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/modelsswagger.Protocol"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Protocol"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "Missing"
                        }
                    }
                }
            }
        },
        "/api/v1/protocols": {
            "get": {
                "description": "Get all protocols",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Protocols"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Protocol"
                        }
                    },
                    "401": {
                        "description": "Request not authorized"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "No"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a protocol",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Protocols"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Add protocol",
                        "name": "type",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/modelsswagger.Protocol"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Protocol"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "Missing"
                        }
                    },
                    "401": {
                        "description": "Request not authorized"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Protocol": {
            "type": "object",
            "properties": {
                "alias": {
                    "type": "string"
                },
                "config": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "md5": {
                    "type": "string"
                },
                "protocol": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
...
        "modelsswagger.Protocol": {
            "type": "object",
            "properties": {
                "protocol": {
                    "type": "string"
                }
            }
        },
...
    }
}